# Use stored alignments if available