Here's the speed up (time saved) for different options. 15.5% time saved across the board and about 18-20% improvement in compute speed.

714s -> 600s. (disk/ssd, filesystem)
633s -> 533s (memory 4k pages, anonymous)
629s -> 533s (memory 2m pages, anonymous)
629s ->  524s (memory 2m pages, filesystem)

333s -> 282s (memory 2m pages, filesystem, alignMethodFor*Graph 3)