@paoloczi 2020-04-09
The comments I made are cosmetic and optional, so I am merging the PR as is. You may want to consider making those changes at a later time.