@bagashe 2020-04-09
Cool. I will make those changes some time today. Excited about a full assembly. Yay!